rollbak in daily price and change from integer to decimal

+++ db/migrate/20210919173016_add_fields_to_property.rb
@@ -3,6 +3,6 @@ class AddFieldsToProperty < ActiveRecord::Migration[6.1]
     add_column :properties, :bathrooms, :integer
     add_column :properties, :parking_spots, :boolean
     add_column :properties, :pets, :boolean
-    add_column :properties, :daily_price, :integer
+    add_column :properties, :daily_price, :decimal
   end
 end
